test case re-org
